requirements.txt
flask
gunicorn
requests
google-cloud-storage

